Fix how metadata is copied on an ABI key match in `AbstractCachingBuildRule`.
Summary:
Previously, we were manually copying the metadata for ABI as well as linear
alloc size. There were no guarantees made about copying _all_ the metadata.
This makes it so that we never miss copying any metadata.

Test Plan: Unit tests.

+++ src/com/facebook/buck/rules/BuildInfoRecorder.java
@@ -100,8 +100,10 @@ public class BuildInfoRecorder {
    * Writes the metadata currently stored in memory to the directory returned by
    * {@link BuildInfo#getPathToMetadataDirectory(BuildTarget)}.
    */
-  public void writeMetadataToDisk() throws IOException {
-    projectFilesystem.rmdir(pathToMetadataDirectory);
+  public void writeMetadataToDisk(boolean clearExistingMetadata) throws IOException {
+    if (clearExistingMetadata) {
+      projectFilesystem.rmdir(pathToMetadataDirectory);
+    }
     projectFilesystem.mkdirs(pathToMetadataDirectory);
 
     for (Map.Entry<String, String> entry : metadataToWrite.entrySet()) {
@@ -112,7 +114,7 @@ public class BuildInfoRecorder {
   }
 
   /**
-   * This key/value pair is stored in memory until {@link #writeMetadataToDisk()} is invoked.
+   * This key/value pair is stored in memory until {@link #writeMetadataToDisk(boolean)} is invoked.
    */
   public void addMetadata(String key, String value) {
     metadataToWrite.put(Preconditions.checkNotNull(key), Preconditions.checkNotNull(value));
